[RS][DDCE-1198] Summary page content changes: nominee.
