+++ test/viewModels/nominees/NomineeDetailsSummaryHelperSpec.scala
@@ -20,13 +20,17 @@ import assets.messages.BaseMessages
 import base.SpecBase
 import controllers.nominees.{routes => nomineesRoutes}
 import models.{CheckMode, UserAnswers}
-import pages.nominees.IsAuthoriseNomineePage
+import pages.nominees.{ChooseNomineePage, IsAuthoriseNomineePage}
 import viewmodels.SummaryListRowHelper
 import viewmodels.nominees.NomineeDetailsSummaryHelper
 
 class NomineeDetailsSummaryHelperSpec extends SpecBase with SummaryListRowHelper {
 
-  private val helper = new NomineeDetailsSummaryHelper(UserAnswers("id")
+  private val helperWithOrganisation = new NomineeDetailsSummaryHelper(UserAnswers("id")
+    .set(IsAuthoriseNomineePage, false)
+    .flatMap(_.set(ChooseNomineePage, false)).success.value)
+
+  private val helperNoNominee = new NomineeDetailsSummaryHelper(UserAnswers("id")
     .set(IsAuthoriseNomineePage, false).success.value)
 
 
@@ -35,11 +39,23 @@ class NomineeDetailsSummaryHelperSpec extends SpecBase with SummaryListRowHelper
     "For the is authorise nominee answer" must {
 
       "have a correctly formatted summary list row" in {
-        helper.authoriseNomineeRow mustBe Some(summaryListRow(
+        helperNoNominee.authoriseNomineeRow mustBe Some(summaryListRow(
           messages("isAuthoriseNominee.checkYourAnswersLabel"),
           messages("site.no"),
           Some(messages("isAuthoriseNominee.checkYourAnswersLabel")),
           nomineesRoutes.IsAuthoriseNomineeController.onPageLoad(CheckMode) -> BaseMessages.changeLink
+        ))
+      }
+    }
+
+    "For the nominee type answer" must {
+
+      "have a correctly formatted summary list row" in {
+        helperWithOrganisation.nomineeTypeRow mustBe Some(summaryListRow(
+          messages("chooseNominee.checkYourAnswersLabel"),
+          messages("chooseNominee.false"),
+          Some(messages("chooseNominee.checkYourAnswersLabel")),
+          nomineesRoutes.ChooseNomineeController.onPageLoad(CheckMode) -> BaseMessages.changeLink
         ))
       }
     }
